feat: robot becomes robotToken
backward compat
