+++ src/index.ts
@@ -9,6 +9,7 @@ export * from './definers/datasets.js'
 export * from './definers/functions.js'
 export * from './definers/resources.js'
 export * from './definers/robots.js'
+export * from './definers/robotTokens.js'
 export * from './definers/roles.js'
 export * from './definers/webhooks.js'
 
@@ -28,6 +29,7 @@ export * from './types/errors.js'
 export * from './types/functions.js'
 export * from './types/resources.js'
 export * from './types/robots.js'
+export * from './types/robotTokens.js'
 export * from './types/roles.js'
 export * from './types/webhooks.js'
 
@@ -42,6 +44,7 @@ export * from './validation/datasets.js'
 export * from './validation/functions.js'
 export * from './validation/resources.js'
 export * from './validation/robots.js'
+export * from './validation/robotTokens.js'
 export * from './validation/roles.js'
 export * from './validation/webhooks.js'
