+++ src/types/robotTokens.ts
@@ -1,0 +1,44 @@
+import type {BlueprintProjectResourceLifecycle, BlueprintResource} from './resources'
+
+/**
+ * Resource types that robot tokens can be attached to.
+ * @internal
+ */
+export type RobotTokenResourceType = 'organization' | 'project'
+
+/**
+ * Defines the robot token's roles within a given resource.
+ * @beta This feature is subject to breaking changes.
+ * @category Resource Types
+ */
+export interface RobotTokenMembership {
+  resourceType: RobotTokenResourceType
+  resourceId: string
+  roleNames: string[]
+}
+
+/**
+ * A robot token that provides automated access.
+ * @see https://www.sanity.io/docs/content-lake/http-auth#k4c21d7b829fe
+ * @beta This feature is subject to breaking changes.
+ * @category Resource Types
+ */
+export interface BlueprintRobotTokenResource extends BlueprintResource<BlueprintProjectResourceLifecycle> {
+  type: 'sanity.access.robot'
+  /** A descriptive label for the robot token and its use case */
+  label: string
+  memberships: RobotTokenMembership[]
+
+  resourceType?: RobotTokenResourceType
+  resourceId?: string
+}
+
+/**
+ * Configuration for a robot token that provides automated access.
+ * @see https://www.sanity.io/docs/content-lake/http-auth#k4c21d7b829fe
+ * @beta This feature is subject to breaking changes.
+ * @category Resource Types
+ */
+export type BlueprintRobotTokenConfig = Omit<BlueprintRobotTokenResource, 'type' | 'label' | 'token'> & {
+  label?: string
+}
